Moved utils to subfolder
