+++ lib/rollup.config.js
@@ -36,14 +36,14 @@ const createUtilsConfigs = ({ name, input }) => {
     typings: 'index.d.ts',
   };
   fs.writeFileSync(
-    `build/${name}/package.json`,
+    `build/utils/${name}/package.json`,
     JSON.stringify(utilsPkg, null, 2)
   );
   return {
-    input: `src/${name}/index.ts`,
+    input: `src/utils/${name}/index.ts`,
     external,
     output: {
-      file: `build/${name}/index.esm.js`,
+      file: `build/utils/${name}/index.esm.js`,
       format: 'esm',
     },
     plugins: [
