+++ lib/src/wrappers/InlineWrapper.tsx
@@ -81,6 +81,7 @@ export class InlineWrapper extends React.PureComponent<
 
   public close = () => {
     this.setState({ anchorEl: null });
+    this.props.handleAccept();
     if (this.props.onClose) {
       this.props.onClose();
     }
